+++ e2e/test-env/src/client/RicosTestApp/editorPlugins.js
@@ -30,6 +30,7 @@ import {
 } from 'wix-rich-content-plugin-vertical-embed';
 import { mockFetchUrlPreviewData } from '../../../../../examples/main/shared/utils/linkPreviewUtil';
 import { pluginTextColor, pluginTextHighlight } from 'wix-rich-content-plugin-text-color';
+import { pluginUnsupportedBlocks } from 'wix-rich-content-plugin-unsupported-blocks';
 
 import { createPresets } from './utils';
 import {
@@ -101,6 +102,7 @@ const createPlugins = externalConfigs => {
   const configs = normalizeConfigs(merge(defaultConfigs, externalConfigs));
 
   return {
+    unsupportedBlocks: pluginUnsupportedBlocks(),
     image: pluginImage(),
     gallery: pluginGallery(configs.gallery),
     video: pluginVideo(configs.video),