+++ packages/editor/web/src/RichContentEditor/createPlugins.ts
@@ -60,6 +60,7 @@ const createPlugins = ({
     isMobile: context.isMobile,
   });
 
+  const supportedBlockTypes = Object.entries(context.config).map(([key]) => key);
   const listPlugin = createListPlugin({ olRegex: /1\./, allowNestedLists: false, ulChars: [] });
 
   const dndPlugin = createBlockDndPlugin();
@@ -81,6 +82,7 @@ const createPlugins = ({
     decorator: wixPluginsDecorators,
     commonPubsub,
     pluginDefaults,
+    supportedBlockTypes,
     ...context,
     ...context.config,
   };