+++ packages/plugin-commons/web/src/Base/createBasePlugin.ts
@@ -29,6 +29,7 @@ import {
   UnderlyingPlugin,
 } from 'wix-rich-content-common';
 import { CSSProperties, ComponentType } from 'react';
+import { UNSUPPORTED_BLOCKS_TYPE } from '../consts';
 
 type EditorStateFuncs = { getEditorState: GetEditorState; setEditorState: SetEditorState };
 
@@ -62,6 +63,7 @@ interface CreateBasePluginConfig extends CreatePluginConfig {
   onOverlayClick?: ({ e, pubsub }: { e: Event; pubsub: Pubsub }) => void;
   onComponentMount?: ({ e, pubsub }: { e: Event; pubsub: Pubsub }) => void;
   disableRightClick?: UISettings['disableRightClick'];
+  supportedBlockTypes?: string[];
   type: PluginType;
   defaultPluginData: Record<string, unknown>;
   decoratorTrigger?: string;
@@ -256,21 +258,33 @@ const createBasePlugin = (
       // TODO subject to change for draft-js next release
       const contentState = getEditorState().getCurrentContent();
       const key = contentBlock.getEntityAt(0);
+
       if (key) {
         const entity = contentState.getEntity(key);
-        const type = entity.getType();
-        if (config.type === type || config.legacyType === type) {
-          return {
-            component: DecoratedCompWithBase,
-            editable: false,
-            props: {
-              getData: getData(contentBlock, { getEditorState }),
-              setData: setData(contentBlock, { getEditorState, setEditorState }),
-              deleteBlock: deleteEntity(contentBlock, { getEditorState, setEditorState }),
-              type: config.type,
-            },
-          };
+        const entityType = entity.getType();
+        let type;
+        if (config.type === entityType || config.legacyType === entityType) {
+          type = entityType;
+        } else if (
+          config.type === UNSUPPORTED_BLOCKS_TYPE &&
+          !config.supportedBlockTypes?.includes(entityType)
+        ) {
+          type = UNSUPPORTED_BLOCKS_TYPE;
         }
+        const blockRenderObject = type
+          ? {
+              component: DecoratedCompWithBase,
+              editable: false,
+              props: {
+                getData: getData(contentBlock, { getEditorState }),
+                setData: setData(contentBlock, { getEditorState, setEditorState }),
+                deleteBlock: deleteEntity(contentBlock, { getEditorState, setEditorState }),
+                type,
+                unsupportedType: type === UNSUPPORTED_BLOCKS_TYPE ? entityType : null,
+              },
+            }
+          : null;
+        return blockRenderObject;
       }
     }
     return null;