+++ packages/plugin-unsupported-blocks/web/src/toolbar/createToolbar.ts
@@ -1,0 +1,9 @@
+import { CreatePluginToolbar } from 'wix-rich-content-common';
+
+const createToolbar: CreatePluginToolbar = () => {
+  return {
+    name: 'unsupported-blocks',
+  };
+};
+
+export default createToolbar;