+++ packages/plugin-unsupported-blocks/web/src/unsupported-blocks-viewer.tsx
@@ -1,0 +1,40 @@
+import React, { Component } from 'react';
+import {
+  mergeStyles,
+  RichContentTheme,
+  ComponentData,
+  TranslationFunction,
+} from 'wix-rich-content-common';
+import styles from '../statics/styles/unsupported-blocks.scss';
+import CircleInfoIcon from './icons/CircleInfoIcon';
+import { UnsupportedBlocksPluginEditorConfig } from './types';
+
+interface Props {
+  componentData: ComponentData;
+  settings: UnsupportedBlocksPluginEditorConfig;
+  theme: RichContentTheme;
+  unsupportedType: string;
+  t: TranslationFunction;
+}
+
+class UnsupportedBlocksViewer extends Component<Props> {
+  styles: Record<string, string>;
+
+  render() {
+    const { unsupportedType, t, theme } = this.props;
+    const pluginName = t(unsupportedType);
+    const unsupportedMessage = t('unsupported_plugin_text');
+    this.styles = this.styles || mergeStyles({ styles, theme });
+
+    return (
+      <div className={styles.unsupportedBlocks_alert}>
+        <CircleInfoIcon />
+        <p>
+          {pluginName} {unsupportedMessage}
+        </p>
+      </div>
+    );
+  }
+}
+
+export default UnsupportedBlocksViewer;